Move path handing out of gui folder
