+++ scripts/Engineering/gui/engineering_diffraction/settings/settings_model.py
@@ -5,7 +5,7 @@
 #   Institut Laue - Langevin & CSNS, Institute of High Energy Physics, CAS
 # SPDX - License - Identifier: GPL - 3.0 +
 from Engineering.gui.engineering_diffraction.settings.settings_helper import get_setting, set_setting
-from Engineering.gui.engineering_diffraction.tabs.common import path_handling
+from Engineering.common import path_handling
 
 
 class SettingsModel(object):
